==> combine_effort/0918. 3-may-2020 to predict 10-may-2020/2. new correlation.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import RandomForestRegressor
from sklearn.neighbors import KNeighborsClassifier
from sklearn.neural_network import MLPClassifier
#import matplotlib.pyplot as plt
from sklearn import metrics
import numpy as np
import time
from datetime import datetime

q = 0
last_trial_time = datetime.now()
while q < 1:
    
    print(last_trial_time)
    control_dataframe = pd.read_csv('control.csv' ) 
    
    if control_dataframe['classify'].iloc[:1].values == 1:     
        
        q = q+1
        
        
        
    else:  
        last_trial_time = datetime.now()
        time.sleep(60)
        
    
    


currency_pairs = ['USDCHF10080',
                 'GBPUSD10080', 'EURUSD10080', 'USDJPY10080', 'USDCAD10080', 'AUDUSD10080', 'NZDUSD10080',
                 'GBPCHF10080', 'EURCHF10080', 'CHFJPY10080', 'CADCHF10080', 'AUDCHF10080', 'NZDCHF10080', 'EURGBP10080',
                 'GBPJPY10080', 'GBPCAD10080', 'GBPAUD10080', 'EURJPY10080', 'EURCAD10080', 'EURAUD10080', 'EURNZD10080',
                'CADJPY10080', 'AUDJPY10080', 'NZDJPY10080', 'AUDCAD10080', 'NZDCAD10080', 'AUDNZD10080'
                ]

headers = ['date', 'ignore', 'open', 'high', 'low', 'close', 'volume'  ]
print('---------------- Script Has Started Execution-----------------------')

# declare the variable to hold all corelation data
combine_correlation_data = pd.DataFrame()

#loop through all currency
for z in currency_pairs:
    
    current_currency = z.replace('10080','')
    raw_data = pd.read_csv('files/' + current_currency + '.csv'  ) 
    
    
    weeks = [50, 30, 20, 15, 10]
    
    #loop through all availale weeks
    for y in weeks:
        
        currency_list = []
        correlation_score = []
        
        #loops through all currency pairs again to calculate the corelation of each currency pair with the current iteration of z
        for x in currency_pairs:          

            #duplicates the raw data to work with
            current_data = raw_data.copy()
            
            #returns on the the amount of rows of weeks needed to perform correlation on
            current_data = current_data.head(y)
            
            #returns the two columns that correlation is being calculated for
            new_df = current_data[['nextweek_'+ current_currency, x.replace('10080','')]]

            correlation = new_df.corr()
            correlation = correlation[x.replace('10080','')][0:1].values[0]
            
            #gets the absolute value of the correlation to 5 digit number
            correlation = abs(round(correlation, 5))
            
            #apppends the current currency and the corelation values to their respective lists
            currency_list.append(x.replace('10080',''))
            correlation_score.append(correlation)
            
            #creates a mini data frame with the information
            current_correlation_dataframe = pd.DataFrame()  
            current_correlation_dataframe [current_currency + '__' + str(y) ] = currency_list
            current_correlation_dataframe [current_currency + '_' + str(y) ] = correlation_score
            
            # sort the dataframe with corelation value from large to small
            current_correlation_dataframe.sort_values(by=[current_currency + '_' + str(y)], ascending=False, inplace=True, ignore_index= True)
        
            #fill the NaN values with zero
            current_correlation_dataframe.fillna(0, inplace=True)
            
        # add the current_currency dataframe to the global dataframe
        combine_correlation_data = pd.concat((combine_correlation_data, current_correlation_dataframe),axis=1)
        #print(len(correlation_score))
    
print(combine_correlation_data)


# save global dataframe to csv
combine_correlation_data.to_csv('files/auto-correlation2.csv', index=False)


control_dataframe = pd.read_csv('control.csv' ) 
control_dataframe['corelation'].iloc[:1] = 1


print(control_dataframe)
control_dataframe.to_csv('control.csv', index=False)
